Move FixAcceptHeader middleware to middlewares subdirectory
