+++ WcaOnRails/config/application.rb
@@ -1,5 +1,5 @@
 require File.expand_path('../boot', __FILE__)
-require_relative '../lib/fix_accept_header'
+require_relative '../lib/middlewares/fix_accept_header'
 require_relative '../lib/middlewares/warden_user_logger'
 
 require 'rails/all'
